Improve build script error handling
Add checks for uncommitted changes and blogdown build success.
The update introduces a precondition check that halts the script and notifies the user
when there are uncommitted or unpushed changes in the git repository.
It utilizes a conditional to assess the presence of changes by using git status.
Additionally, the script now tests the success of the blogdown::build_site command
and terminates the script if the build process fails.
This enhancement aims to prevent incomplete or incorrect builds due to pending changes
or failed R markdown builds.

+++ build.sh
@@ -1,13 +1,18 @@
-set -e
+if [[ $(git status --porcelain) ]]; then
+    echo "There are uncommitted or unpushed changes. Please commit or push your changes before running this script."
+    exit 1
+fi
+
 pip install \
     matplotlib \
     numpy \
-    pandas 
+    pandas
 
-r -e 'blogdown::build_site(build_rmd = TRUE)'
+if ! R -e 'blogdown::build_site(build_rmd = TRUE)'; then
+    exit 1
+fi
 
 git add public
 git commit --amend --no-edit
 git push --force-with-lease
-
